Re-add text type

+++ src/storage_node/metadata/field_type.go
@@ -74,7 +74,7 @@ func (f *FieldType) Normalize(val interface{}) (interface{}, error) {
     }],
     "datasource_field_type": {
         "1": {
-            "type": "int",
+            "type": "smallint",
             "args": {
                 "size": 2
             }
